added Chat Messaging functionality
