--- app/src/main/java/com/example/space__chat/data/model/ChatDatabase.kt
@@ -1,14 +1,0 @@
-package com.example.space__chat.data
-
-import androidx.room.Database
-import androidx.room.RoomDatabase
-import com.example.space__chat.data.model.ChatEntity
-
-@Database(
-    entities = [ChatEntity::class],
-    version = 2,
-)
-abstract class ChatDataBase : RoomDatabase() {
-
-    abstract fun chatDao(): ChatDao
-}
