--- app/src/main/java/com/example/space__chat/data/model/MessageModel.kt
@@ -1,10 +1,0 @@
-package com.example.space__chat.data.model
-
-import com.example.space__chat.presentation.model.UserType
-
-data class MessageModel(
-    val id: Int?,
-    val sender: UserType?,
-    val message: String?,
-    val timeStamp: Long?
-)
