--- app/src/main/java/com/example/space__chat/di/DatabaseModul.kt
@@ -1,17 +1,0 @@
-package com.example.space__chat.di
-
-import android.content.Context
-import androidx.room.Room
-import com.example.space__chat.data.ChatDataBase
-import org.koin.dsl.module
-
-private fun provideChatDB(context: Context): ChatDataBase {
-    return Room.databaseBuilder(context, ChatDataBase::class.java, "database_1")
-        .fallbackToDestructiveMigration().build()
-}
-private fun provideDao(db: ChatDataBase) = db.chatDao()
-
-val dbModule = module {
-    single { provideChatDB(get()) }
-    single { provideDao(get()) }
-}
