--- app/src/main/java/com/example/space__chat/domain/ChatRepository.kt
@@ -1,9 +1,0 @@
-package com.example.space__chat.domain
-
-import com.example.space__chat.data.model.MessageModel
-import kotlinx.coroutines.flow.Flow
-
-interface ChatRepository {
-    fun getMessages(): Flow<List<MessageModel>>
-    suspend fun insertMessages(message: MessageModel)
-}
