--- app/src/main/java/com/example/space__chat/presentation/model/ThemeMode.kt
@@ -1,6 +1,0 @@
-package com.example.space__chat.presentation.model
-
-enum class ThemeMode {
-    LIGHT,
-    DARK
-}
