--- app/src/main/java/com/example/space__chat/presentation/model/UserType.kt
@@ -1,6 +1,0 @@
-package com.example.space__chat.presentation.model
-
-enum class UserType {
-    TOP_USER,
-    BOTTOM_USER
-}
