--- app/src/main/java/com/example/space__chat/utils/DiffUtil.kt
@@ -1,16 +1,0 @@
-package com.example.space__chat.utils
-
-import android.annotation.SuppressLint
-import androidx.recyclerview.widget.DiffUtil
-
-class ChatCallBack<T : Any> : DiffUtil.ItemCallback<T>() {
-
-    override fun areItemsTheSame(oldItem: T, newItem: T): Boolean {
-        return oldItem === newItem
-    }
-
-    @SuppressLint("DiffUtilEquals")
-    override fun areContentsTheSame(oldItem: T, newItem: T): Boolean {
-        return oldItem == newItem
-    }
-}
