+++ app/src/main/java/com/space/space__chat/data/ChatDao.kt
@@ -1,0 +1,16 @@
+package com.space.space__chat.data
+
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.Query
+import com.space.space__chat.data.model.ChatEntity
+import kotlinx.coroutines.flow.Flow
+
+@Dao
+interface ChatDao {
+    @Query("SELECT * FROM messages")
+    fun getAllMessages(): Flow<List<ChatEntity>>
+
+    @Insert
+    suspend fun insertMessages(vararg message : ChatEntity)
+}
