+++ app/src/main/java/com/space/space__chat/data/mapper/Mapper.kt
@@ -1,0 +1,17 @@
+package com.space.space__chat.data.mapper
+
+import com.space.space__chat.data.model.ChatEntity
+import com.space.space__chat.data.model.MessageModel
+
+fun MessageModel.toEntity() = ChatEntity(
+    id = id,
+    sender = sender,
+    message = message,
+    timeStamp = timeStamp,
+)
+fun ChatEntity.toModel() = MessageModel(
+    id = id,
+    sender = sender,
+    message = message,
+    timeStamp = timeStamp,
+)
