+++ app/src/main/java/com/space/space__chat/data/model/ChatEntity.kt
@@ -1,0 +1,14 @@
+package com.space.space__chat.data.model
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+import com.space.space__chat.presentation.model.UserType
+
+@Entity(tableName = "messages")
+data class ChatEntity(
+    @PrimaryKey(autoGenerate = true)
+    val id: Int? =null,
+    val sender: UserType?,
+    val message: String?,
+    val timeStamp: Long?
+)
