+++ app/src/main/java/com/space/space__chat/domain/ChatRepository.kt
@@ -1,0 +1,9 @@
+package com.space.space__chat.domain
+
+import com.space.space__chat.data.model.MessageModel
+import kotlinx.coroutines.flow.Flow
+
+interface ChatRepository {
+    fun getMessages(): Flow<List<MessageModel>>
+    suspend fun insertMessages(message: MessageModel)
+}
