+++ app/src/main/java/com/space/space__chat/presentation/model/ThemeMode.kt
@@ -1,0 +1,6 @@
+package com.space.space__chat.presentation.model
+
+enum class ThemeMode {
+    LIGHT,
+    DARK
+}
