+++ app/src/main/java/com/space/space__chat/presentation/model/UserType.kt
@@ -1,0 +1,6 @@
+package com.space.space__chat.presentation.model
+
+enum class UserType {
+    TOP_USER,
+    BOTTOM_USER
+}
