+++ app/src/main/java/com/space/space__chat/utils/extensions/DateExtension.kt
@@ -1,0 +1,13 @@
+package com.space.space__chat.utils.extensions
+
+import java.text.SimpleDateFormat
+import java.util.*
+
+fun Long.convertTimeToString(): String {
+    val calendar = Calendar.getInstance()
+    val dayMonthFormat = SimpleDateFormat("dd/MM, HH:mm", Locale.getDefault())
+    return dayMonthFormat.format(calendar.time)
+}
+fun getTimeInMills(): Long{
+    return System.currentTimeMillis()
+}
