+++ app/src/main/java/com/space/space__chat/utils/extensions/ViewExtensions.kt
@@ -1,0 +1,14 @@
+package com.space.space__chat.utils.extensions
+
+import android.content.res.ColorStateList
+import android.view.View
+import android.widget.ImageView
+
+
+fun View.setTint(colorRes: Int) {
+    this.backgroundTintList = ColorStateList.valueOf(this.context.getColor(colorRes))
+}
+
+fun ImageView.setImgTint(colorRes: Int) {
+    this.imageTintList = ColorStateList.valueOf(this.context.getColor(colorRes))
+}
