+++ app/src/main/res/values-night/themes.xml
@@ -13,6 +13,6 @@
         <item name="android:statusBarColor">?attr/colorPrimaryVariant</item>
         <!-- Customize your theme here. -->
         <item name="android:background">@color/dark_01_midnight_blue</item>
-        <item name="android:textColor">@color/neutral_06_white</item>
+        <item name="color">@color/neutral_06_white</item>
     </style>
 </resources>
