+++ app/src/main/res/values/text_style.xml
@@ -2,7 +2,7 @@
 <resources>
     <style name="mediumMyriadGeo_Normal">
         <item name="android:textSize">14sp</item>
-        <item name="android:textColor">@color/neutral_01_great_dark_grey</item>
+        <item name="android:textColor">?attr/color</item>
         <item name="android:font">@font/myriad_geo</item>
     </style>
