+++ app/src/main/res/values/themes.xml
@@ -13,6 +13,6 @@
         <item name="android:statusBarColor">?attr/colorPrimaryVariant</item>
         <!-- Customize your theme here. -->
         <item name="android:background">@color/neutral_06_white</item>
-        <item name="android:textColor">@color/black</item>
+        <item name="color">@color/black</item>
     </style>
 </resources>
